BackEnd/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
langchain-core
openai==1.10.0
supabase==2.3.4
python-dotenv==1.0.0
httpx>=0.24.0,<0.26.0